test(Tabs): rename SVM edge-effects scenario to scroll-edge-effects
Clarify that the scenario covers scroll edge effects specifically.
Renames the directory, scenario key, and import symbol.
