+++ apps/src/tests/component-integration-tests/scroll-view-marker/index.ts
@@ -1,8 +1,11 @@
 import type { ScenarioGroup } from '@apps/tests/shared/helpers';
-import TestSvmTabsEdgeEffects from './test-svm-tabs-edge-effects';
+import TestSvmTabsScrollEdgeEffects from './test-svm-tabs-scroll-edge-effects';
 import TestStackSvmTabsSpecialEffects from './test-stack-svm-tabs-special-effects';
 
-const scenarios = { TestSvmTabsEdgeEffects, TestStackSvmTabsSpecialEffects };
+const scenarios = {
+  TestSvmTabsScrollEdgeEffects,
+  TestStackSvmTabsSpecialEffects,
+};
 
 const TestSvmScenarioGroup: ScenarioGroup<keyof typeof scenarios> = {
   name: 'ScrollViewMarker integration cases',
