+++ apps/src/tests/component-integration-tests/scroll-view-marker/test-svm-tabs-scroll-edge-effects/scenario-description.ts
@@ -2,7 +2,7 @@ import type { ScenarioDescription } from '@apps/tests/shared/helpers';
 
 export const scenarioDescription: ScenarioDescription = {
   name: 'Integration: SVM in Tabs - edge effects',
-  key: 'test-svm-tabs-edge-effects',
+  key: 'test-svm-tabs-scroll-edge-effects',
   details:
     'Test whether scroll edge effects are applied correctly when ScrollViewMarker ' +
     'is used inside Tabs',
